Add CSS files for components and restructure
