+++ src/components/BillsList/BillsList.css
@@ -1,0 +1,1 @@
+/* Styles for BillsList */
