+++ src/components/BillsList/DashboardBillsList.css
@@ -1,0 +1,1 @@
+/* Styles for DashboardBillsList */
