+++ src/components/ChartsPage/ChartsPage.css
@@ -1,0 +1,1 @@
+/* Styles for ChartsPage */
