+++ src/components/ChartsPage/CreditUtilizationGauge.css
@@ -1,0 +1,1 @@
+/* Styles for CreditUtilizationGauge */
