+++ src/components/ChartsPage/ExpenseBreakdownChart.css
@@ -1,0 +1,1 @@
+/* Styles for ExpenseBreakdownChart */
