+++ src/components/ChartsPage/ExpenseTypeChart.css
@@ -1,0 +1,1 @@
+/* Styles for ExpenseTypeChart */
