+++ src/components/ErrorBoundary/ErrorBoundary.css
@@ -1,0 +1,1 @@
+/* Styles for ErrorBoundary */
