+++ src/components/ErrorBoundary/ErrorBoundary.jsx
@@ -1,6 +1,7 @@
 // src/components/ErrorBoundary.jsx
 import React from 'react';
 import { Result, Button } from 'antd';
+import './ErrorBoundary.css';
 
 class ErrorBoundary extends React.Component {
   constructor(props) {
