+++ src/components/FinancialSummary/CombinedBillsOverview/CombinedBillsOverview.jsx
@@ -27,6 +27,7 @@ import updateLocale from 'dayjs/plugin/updateLocale';
 import isSameOrBefore from 'dayjs/plugin/isSameOrBefore';
 import isSameOrAfter from 'dayjs/plugin/isSameOrAfter';
 import isBetween from 'dayjs/plugin/isBetween';
+import './CombinedBillsOverview.css';
 
 // Extend dayjs plugins
 dayjs.extend(relativeTime);
