+++ src/components/FinancialSummary/CombinedBillsOverview/MonthlyProgressSummary.css
@@ -1,0 +1,1 @@
+/* Styles for MonthlyProgressSummary */
