+++ src/components/FinancialSummary/MonthlyProgressCard/MonthlyProgressCard.css
@@ -1,0 +1,1 @@
+/* Styles for MonthlyProgressCard */
