+++ src/components/FinancialSummary/MonthlyProgressCard/MonthlyProgressCard.jsx
@@ -9,6 +9,7 @@ import {
     IconCurrencyDollar // Icon for Total Amount
 } from '@tabler/icons-react';
 import { FinanceContext } from '../../contexts/FinanceContext';
+import './MonthlyProgressCard.css';
 
 const { Text, Title } = Typography;
