+++ src/components/Navigation/BottomNavBar/BottomNavBar.css
@@ -1,0 +1,1 @@
+/* Styles for BottomNavBar */
