+++ src/components/RecentActivity/ActivityFeed.css
@@ -1,0 +1,1 @@
+/* Styles for ActivityFeed */
