+++ src/components/RecentActivity/NonRecurringTransactions.css
@@ -1,0 +1,1 @@
+/* Styles for NonRecurringTransactions */
