+++ src/components/RecentActivity/PastDuePayments.css
@@ -1,0 +1,1 @@
+/* Styles for PastDuePayments */
