+++ src/components/RecentActivity/UpcomingPayments.css
@@ -1,0 +1,1 @@
+/* Styles for UpcomingPayments */
