+++ src/components/shared/CardLayout/CardLayout.css
@@ -1,0 +1,1 @@
+/* Styles for CardLayout */
