+++ src/components/skeletons/FinancialOverviewSkeleton/FinancialOverviewSkeleton.css
@@ -1,0 +1,1 @@
+/* Styles for FinancialOverviewSkeleton */
